+++ Changelog.md
@@ -7,6 +7,7 @@
 
 - Added the Adjustments module
 - Added the extended Adjustable Cart model to `Foundation` (original Cart model still there, intact)
+- Added Shipping methods
 - Added `actives()` and `inactives()` scopes to the Carrier model
 - Changed minimum Konekt module requirements to:
     - Address: 2.2